requirements.txt
kafka-python==2.0.2
confluent-kafka==2.3.0
faker==22.0.0
pymongo==4.6.1

